Port numbers are limited to 16 bit.
git-svn-id: file:///home/asfernandes/fb/repository/sf/svn/firebird/trunk@11011 de594faa-8d1b-4a0c-9a6a-a7de5f8bf859

+++ src/common/config/config.cpp
@@ -395,9 +395,9 @@ const char *Config::getRemoteServiceName()
 	return (const char*) sysConfig.values[KEY_REMOTE_SERVICE_NAME];
 }
 
-int Config::getRemoteServicePort()
+unsigned short Config::getRemoteServicePort()
 {
-	return (int) sysConfig.values[KEY_REMOTE_SERVICE_PORT];
+	return (unsigned short) sysConfig.values[KEY_REMOTE_SERVICE_PORT];
 }
 
 const char *Config::getRemotePipeName()
